Split battle.ts into board, combatant, strike, and engine modules
battle.ts had grown to ~980 lines mixing four distinct concerns: battle
board/hazard data, the BattleCreature combatant, strike/phase resolution
types, and the Battle engine class that ties them together. Splitting
them into src/models/battle/{board,combatant,strike,engine}.ts makes
each concern independently readable and testable.

battle.ts now re-exports the original public surface from these modules
so every existing import site keeps working unchanged. No logic changes.
